Ensure every line from phase output has prefix
